Merge main development branch

+++ seispy/burrow.py
@@ -1,28 +1,32 @@
-from gazelle.datascope import closing,\
-                              dbopen
+import gazelle.datascope as ds
 import seispy.station
 from seispy.util import validate_time
 from obspy.core import read,\
                        Stream
 import os
+import psutil as ps
 import shutil
 import subprocess
 import tempfile
+import time
+
 
 class Groundhog:
-    def __init__(self, server_directory):
+    def __init__(self,
+                 server_directory="rsync://eqinfo.ucsd.edu/ANZA_waveforms",
+                 database_dir="/home/seismech-00/sjfzdb/anf/wfdiscs"):
         self.server_directory = server_directory
         self.dbs = {}
         self.wfdiscs = {}
         for year in range(1998, 2016):
-            self.dbs[year] = dbopen("/home/seismech-00/sjfzdb/anf/wfdiscs/%d" % year)
+            self.dbs[year] = ds.dbopen(os.path.join(database_dir, str(year)))
         self.temp_dir = tempfile.mkdtemp()
-        print self.temp_dir
 
     def __del__(self):
         for year in self.dbs:
             self.dbs[year].close()
-        shutil.rmtree(self.temp_dir)
+        if hasattr(self, "temp_dir"):
+            shutil.rmtree(self.temp_dir)
 
     def fetch(self, station, channel, starttime, endtime):
         """
@@ -53,6 +57,9 @@ class Groundhog:
                                                               endtime.timestamp))
             for data_record in view_data.iter_record():
                 ddir, dfile = data_record.getv("dir", "dfile")
+                while n_rsync_processes() >= 6:
+                    print "SLEEPING,", n_rsync_processes()
+                    time.sleep(1)
                 with open(os.devnull, 'w') as FNULL:
                     subprocess.call(["rsync",
                                      "-a",
@@ -64,7 +71,24 @@ class Groundhog:
                                                    dfile),
                                      self.temp_dir],
                                      stdout=FNULL)
+                os.listdir(self.temp_dir)
                 st += read(os.path.join(self.temp_dir, dfile))
                 os.remove(os.path.join(self.temp_dir, dfile))
+        if len(st) == 0:
+            raise IOError("data not found")
         st.trim(starttime, endtime)
         return st
+
+def n_rsync_processes():
+    try:
+        n = len([p.parent()
+                 for p in ps.process_iter()
+                 if p.name() == "rsync"]) / 2
+    except ps.NoSuchProcess:
+        return n_rsync_processes()
+    return n
+
+if __name__ == "__main__":
+    willy = Groundhog("rsync://eqinfo.ucsd.edu/ANZA_waveforms")
+    st = willy.fetch("PFO", "HHZ", validate_time("2015-120T00:00:00"), validate_time("2015-120T00:01:00"))
+    st.plot()
